fix(func): clean.js testing

+++ src/commands/Moderation/clean.js
@@ -12,12 +12,12 @@ module.exports = {
     botPerms: 8192,
     guildOnly: true,
     run: async (client, message, args) => {
-        if (!args.length) return client.errEmb('No Amount Specified.\n```\nclean <Amount:Number>\nclean <Amount:Number> [User:Mention/ID]\nclean <Amount:Number> [...Flags]\n```\n**Flags:**\n`-users` - Users only\n`-bots` - Bots Only\n`-has <Word>` - Messages that contain a specific word/char', message);
+        if (!args.length) return client.errEmb('No Amount Specified.\n```\nclean <Amount:Number>\nclean <Amount:Number> [User:Mention/ID]\nclean <Amount:Number> [...Flags]\n```\n**Flags:**\n`-users` - Users only\n`-bots` - Bots only\n`-nopin` - No pins\n`-has <Word>` - Messages that contain a specific word/char\n`-to <Message:ID>` - Messages before that message', message);
         let amount = parseInt(args[0]);
         if (isNaN(amount) || amount < 1 || amount > 100) return client.errEmb('Invalid Amount Specified. Amount must be a number between 1 and 100.', message);
         if (args.length > 1) {
             var flagArgs = args.slice(1);
-            let target = '', flagUsers = false, flagBots = false, flagNopin = false, flagHas = '';
+            let target = '', flagUsers = false, flagBots = false, flagNopin = false, flagHas = '', flagTo = 0;
             for (const arg of flagArgs) {
                 if (/(<@!?)?\d{17,19}>?/g.test(arg)) target = /(?:<@!?)?(\d{17,19})>?/g.exec(arg)[1];
                 if (arg.toLowerCase() === '-users') flagUsers = true;
@@ -29,10 +29,19 @@ module.exports = {
                         flagHas = flagArgs[index+1].trim();
                     }
                 }
+                if (arg.toLowerCase() === '-to') {
+                    const index = flagArgs.indexOf(arg);
+                    if (index < flagArgs.length) {
+                        flagTo = flagArgs[index+1].trim();
+                    }
+                }
             }
+            console.log(flagTo);
             if (flagUsers && flagBots) return client.errEmb('Both Users & Bots Flags Specified, pick one.', message);
+            if (flagTo && /\D+/g.test(flagTo)) return client.errEmb('Invalid Message ID for To Flag.', message);
             try {
-                const res = await advanceClean(message, amount, {target:target, flagUsers:flagUsers, flagBots:flagBots, flagNopin:flagNopin, flagHas:flagHas});
+                await message.delete().catch(()=>{});
+                const res = await advanceClean(message, amount, {target:target, flagUsers:flagUsers, flagBots:flagBots, flagNopin:flagNopin, flagHas:flagHas, flagTo:flagTo});
                 return client.checkEmb(`Deleted \`${res}\` Messages!`, message).then(m => setTimeout(() => m.delete(), 3000));
             } catch (err) {
                 return client.errEmb(err.message, message);
